# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ERP Intelligence Dashboard API",
    "description": "Backend API untuk ERP Intelligence Dashboard ISP. Menyajikan wawasan analitis, model prediksi, dan data terstruktur Feature Store.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ingestion/trigger/{job_name}": {
      "post": {
        "tags": [
          "Ingestion & Pipeline"
        ],
        "summary": "Trigger Ingestion Batch Pipeline Manually",
        "description": "Memicu eksekusi sinkronisasi pipeline batch data ERP secara manual. Job yang tersedia: 'subscription_sync' dan 'billing_sync'. PERHATIAN: Jika sistem berjalan dalam fallback mode, data akan ditandai dengan 'is_mock_data: true'.",
        "operationId": "trigger_job_api_v1_ingestion_trigger__job_name__post",
        "parameters": [
          {
            "name": "job_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/JobTriggerRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_JobTriggerResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingestion/history": {
      "get": {
        "tags": [
          "Ingestion & Pipeline"
        ],
        "summary": "Get ETL Batch Execution Logs",
        "description": "Mengembalikan riwayat batch eksekusi ingestion pipeline dengan indikator mock/live yang jelas.",
        "operationId": "get_history_api_v1_ingestion_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_List_BatchLogItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingestion/quality-reports": {
      "get": {
        "tags": [
          "Ingestion & Pipeline"
        ],
        "summary": "Get Data Quality Audit Logs",
        "description": "Mengembalikan laporan audit pemeriksaan kualitas data (null checks, non-negative checks).",
        "operationId": "get_quality_reports_api_v1_ingestion_quality_reports_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "table_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Table Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_List_DataQualityLogItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Autentikasi"
        ],
        "summary": "Login Pengguna",
        "description": "Menghasilkan access token JWT Bearer untuk mengakses endpoint terlindungi.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_TokenResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Autentikasi"
        ],
        "summary": "Profil Pengguna Aktif",
        "description": "Mengembalikan profil dan permission pengguna dari token JWT aktif.",
        "operationId": "get_profile_api_v1_auth_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_UserResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchLogItem": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "source_table": {
            "type": "string",
            "title": "Source Table"
          },
          "watermark_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watermark Start"
          },
          "watermark_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watermark End"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_mock_data": {
            "type": "boolean",
            "title": "Is Mock Data",
            "description": "True jika log batch dieksekusi dengan mock data generator"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "source_table",
          "status",
          "is_mock_data",
          "row_count",
          "started_at"
        ],
        "title": "BatchLogItem"
      },
...
      "DataQualityLogItem": {
        "properties": {
          "log_id": {
            "type": "integer",
            "title": "Log Id"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "check_name": {
            "type": "string",
            "title": "Check Name"
          },
          "table_name": {
            "type": "string",
            "title": "Table Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "checked_at": {
            "type": "string",
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "log_id",
          "check_name",
          "table_name",
          "status",
          "checked_at"
        ],
        "title": "DataQualityLogItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobTriggerRequest": {
        "properties": {
          "limit": {
            "type": "integer",
            "maximum": 50000.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maksimal jumlah baris data yang diekstrak per batch",
            "default": 5000
          },
          "force_full_refresh": {
            "type": "boolean",
            "title": "Force Full Refresh",
            "description": "Jika true, abaikan watermark dan lakukan refresh menyeluruh",
            "default": false
          }
        },
        "type": "object",
        "title": "JobTriggerRequest"
      },
      "JobTriggerResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id",
            "description": "UUID identifier batch yang dieksekusi"
          },
          "job_name": {
            "type": "string",
            "title": "Job Name",
            "description": "Nama pipeline job yang dijalankan"
          },
          "is_mock_data": {
            "type": "boolean",
            "title": "Is Mock Data",
            "description": "TANDA JELAS: True jika data diproduksi dari Mock ERP Generator, False jika Live ERP"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type",
            "description": "'MOCK_GENERATOR' atau 'LIVE_ERP'"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status akhir batch (SUCCESS / FAILED)"
          },
          "rows_extracted": {
            "type": "integer",
            "title": "Rows Extracted",
            "description": "Total baris data yang diekstrak"
          },
          "rows_staged": {
            "type": "integer",
            "title": "Rows Staged",
            "description": "Total baris data yang dimuat ke staging layer"
          },
          "rows_dimension": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Dimension",
            "description": "Total perubahan dimensi (SCD2)"
          },
          "rows_fact": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Fact",
            "description": "Total fact snapshots/monthly yang tersimpan"
          },
          "rows_features": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Features",
            "description": "Total baris fitur ML yang diperbarui"
          },
          "invoices_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoices Count"
          },
          "payments_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payments Count"
          },
          "watermark_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watermark Start"
          },
          "watermark_end": {
            "type": "string",
            "title": "Watermark End"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "job_name",
          "is_mock_data",
          "source_type",
          "status",
          "rows_extracted",
          "rows_staged",
          "watermark_end"
        ],
        "title": "JobTriggerResponse"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "title": "Username",
            "description": "Username atau email akun pengguna"
          },
          "password": {
            "type": "string",
            "minLength": 4,
            "title": "Password",
            "description": "Kata sandi akun"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "ResponseEnvelope_JobTriggerResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Status keberhasilan operasi"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobTriggerResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data utama respons"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Metadata pendukung"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Objek error jika operasi gagal"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseEnvelope[JobTriggerResponse]"
      },
      "ResponseEnvelope_List_BatchLogItem__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Status keberhasilan operasi"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BatchLogItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Payload data utama respons"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Metadata pendukung"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Objek error jika operasi gagal"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseEnvelope[List[BatchLogItem]]"
      },
...
      "ResponseEnvelope_List_DataQualityLogItem__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Status keberhasilan operasi"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DataQualityLogItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Payload data utama respons"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Metadata pendukung"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Objek error jika operasi gagal"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseEnvelope[List[DataQualityLogItem]]"
      },
...
      "ResponseEnvelope_TokenResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Status keberhasilan operasi"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data utama respons"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Metadata pendukung"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Objek error jika operasi gagal"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseEnvelope[TokenResponse]"
      },
      "ResponseEnvelope_UserResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Status keberhasilan operasi"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data utama respons"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Metadata pendukung"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Objek error jika operasi gagal"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseEnvelope[UserResponse]"
      },
      "ResponseError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Katalog kode error standar (e.g. VALIDATION_ERROR, NOT_FOUND)"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Pesan deskriptif error yang ramah pengguna"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Rincian error per-field"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ResponseError"
      },
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Pengenal unik transaksi request (UUIDv4) untuk tracing"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Waktu pembuatan respons dalam format ISO-8601 UTC"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Versi API yang melayani request",
            "default": "v1"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              },
              {
                "type": "null"
              }
            ],
            "description": "Metadata pagination (jika respons mengembalikan daftar data)"
          }
        },
        "type": "object",
        "title": "ResponseMeta"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT Bearer token untuk autentikasi API"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Tipe token",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Masa berlaku token dalam detik"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse",
            "description": "Informasi profil pengguna terotentikasi"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}